+++ examples/pwm_lptim.rs
@@ -8,6 +8,8 @@ use cortex_m::asm;
 use cortex_m_rt::entry;
 #[macro_use]
 mod utilities;
+
+use embedded_hal::pwm::*;
 use stm32h7xx_hal::{pac, prelude::*, rcc::rec};
 
 use log::info;
@@ -44,26 +46,26 @@ fn main() -> ! {
             .kernel_clk_mux(rec::Lptim2ClkSel::Lsi),
         &ccdr.clocks,
     );
-    pwm.enable(); // must be enabled before use
+    pwm.enable().unwrap(); // must be enabled before use
 
     // Output PWM on PB13
-    let max = pwm.get_max_duty();
-    pwm.set_duty(max / 2);
+    let max = pwm.max_duty_cycle();
+    pwm.set_duty_cycle(max / 2).unwrap();
 
     info!("50%");
-    pwm.enable();
+    pwm.enable().unwrap();
     asm::bkpt();
 
     info!("25%");
-    pwm.set_duty(max / 4);
+    pwm.set_duty_cycle(max / 4).unwrap();
     asm::bkpt();
 
     info!("12.5%");
-    pwm.set_duty(max / 8);
+    pwm.set_duty_cycle(max / 8).unwrap();
     asm::bkpt();
 
     info!("100%");
-    pwm.set_duty(max);
+    pwm.set_duty_cycle(max).unwrap();
     asm::bkpt();
 
     loop {